realestate/requirements.txt
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.22
chromadb==0.4.22
openai==1.3.7
python-dotenv==1.0.0
streamlit==1.29.0
pandas==2.1.4
numpy==1.26.2
beautifulsoup4==4.12.2
requests==2.31.0
PyPDF2==3.0.1
tiktoken==0.5.1
pandas==2.1.4
